results/generals_io_rl_demo/requirements.txt
generals-bots
pygame # Required for 'render_mode="human"' if you enable GUI rendering locally. Not strictly needed for headless Docker execution.